+++ spec/dummy/client/entries/app-kitten.js
@@ -47,6 +47,9 @@ import ExternalRichLink from 'kitten/components/links/external-rich-link'
 // Lists
 import NavList from 'kitten/components/lists/nav-list'
 
+// Meters
+import { Progress } from 'kitten/components/meters/progress'
+
 // Popovers
 import { Popover } from 'kitten/components/popovers/popover'
 import { CallToActionPopover } from 'kitten/components/popovers/call-to-action-popover'
@@ -77,6 +80,8 @@ import { Title } from 'kitten/components/typography/title'
 
 // Karl composed components
 
+import { KarlAutoTriggerPopover }
+  from 'karl/components/popovers/auto-trigger-popover'
 import { KarlLoudspeakerPopover } from 'karl/components/popovers/loudspeaker-popover'
 import { KarlImageExternalRichLink,
          KarlSvgExternalRichLink } from 'kitten/karl/molecules/external-rich-link'
@@ -116,6 +121,7 @@ import { KarlButtonIconHeartRight,
        } from 'kitten/karl/buttons/button-heart-icon-words'
 import { KarlButtonAndText } from 'kitten/karl/buttons/button'
 import { KarlPaymentButton } from 'kitten/karl/buttons/payment-button'
+import { KarlRandomProgress } from 'kitten/karl/meters/random-progress'
 import { KarlMakerWhoAmIStep } from 'app/maker/who-am-i/components/container'
 
 ReactOnRails.register({
@@ -167,6 +173,9 @@ ReactOnRails.register({
   // Lists
   NavList,
 
+  // Meters
+  Progress,
+
   // Popovers
   CallToActionPopover,
   Popover,
@@ -196,6 +205,7 @@ ReactOnRails.register({
   Title,
 
   // Karl composed components
+  KarlAutoTriggerPopover,
   KarlLoudspeakerPopover,
   KarlButtonIconHeartRight,
   KarlButtonIconHeartLeft,
@@ -230,5 +240,6 @@ ReactOnRails.register({
   KarlParagraphQuaternary,
   KarlParagraphWithLink,
   KarlPaymentButton,
+  KarlRandomProgress,
   KarlMakerWhoAmIStep,
 })