+++ CHANGELOG.md
@@ -5,7 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add new `TextInputWithCounter` React component.
-- Feature: Add new `TextInput` React component.
+- Feature: Add new `Progress` React component.
+
+## [5.9.0] - 2017-02-06
+
+Features:
+- Add new `TextInput` React component.
+- Add `autoTriggerEnhancer` high-order component.
 
 ## [5.8.0] - 2017-02-06